+++ fundraiser/tests/functions/mod.rs
@@ -1,0 +1,14 @@
+mod asset_count;
+mod asset_info_by_count;
+mod asset_info_by_id;
+mod campaign;
+mod campaign_info;
+mod cancel_campaign;
+mod claim_pledges;
+mod create_campaign;
+mod pledge;
+mod pledge_count;
+mod pledged;
+mod total_campaigns;
+mod unpledge;
+mod user_campaign_count;
